# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Address:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - country
    - postal_code
    - state
    - street
    type: object
...
  models.CreateOrderRequest:
    properties:
      customer_id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        minItems: 1
        type: array
      shipping_address:
        $ref: '#/definitions/models.Address'
    required:
    - customer_id
    - items
    - shipping_address
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        minItems: 1
        type: array
      payment_intent_id:
        type: string
      payment_status:
        $ref: '#/definitions/models.PaymentStatus'
      shipping_address:
        $ref: '#/definitions/models.Address'
      status:
        $ref: '#/definitions/models.OrderStatus'
      total_amount:
        type: number
      updated_at:
        type: string
    required:
    - customer_id
    - items
    - shipping_address
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
      unit_price:
        minimum: 0
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
  models.OrderStatus:
    enum:
    - pending
    - confirmed
    - shipping
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusConfirmed
    - OrderStatusShipping
    - OrderStatusDelivered
    - OrderStatusCancelled
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  models.PaymentStatus:
    enum:
    - pending
    - succeeded
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusSucceeded
    - PaymentStatusFailed
    - PaymentStatusRefunded
...
  models.UpdateOrderStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.OrderStatus'
        enum:
        - pending
        - confirmed
        - shipping
        - delivered
        - cancelled
    required:
    - status
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: aaravmahajan2003@gmail.com
    name: Aarav Mahajan
    url: https://github.com/aaravmahajanofficial
  description: This is the API server for the Scalable E-commerce Platform. It provides
    endpoints for managing users, products, carts, orders, payments, and notifications.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scalable E-commerce Platform API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Retrieves a paginated list of orders placed by the authenticated
        user. Requires authentication.
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of orders
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Order'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's orders with pagination
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Creates a new order from the user's current cart items and provided
        shipping details. Requires authentication.
      parameters:
      - description: Order Creation Details (includes shipping, uses current cart)
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created order
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Validation error, empty cart, or insufficient stock
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Cart not found (should be created implicitly if needed)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      description: Retrieves details for a specific order placed by the authenticated
        user. Requires authentication.
      parameters:
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved order
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid order ID format
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - User does not own this order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an order by ID
      tags:
      - Orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of a specific order. Requires authentication
        (potentially admin-level).
      parameters:
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: New Order Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated order status
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid order ID format or invalid status value
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - Insufficient permissions to update status" // If
            applicable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update order status (Admin/Internal)
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"